+++ packages/flags/src/index.ts
@@ -1,0 +1,6 @@
+export {
+  /**
+   * Metadata for all available onyx flags.
+   */
+  default as FLAG_METADATA,
+} from "./metadata.json";