+++ packages/sit-onyx/src/components/OnyxDataGrid/examples/SelectionExample.vue
@@ -1,11 +1,6 @@
 <script setup lang="ts">
 import { ref } from "vue";
-import {
-  DataGridFeatures,
-  OnyxDataGrid,
-  type ColumnConfig,
-  type ColumnGroupConfig,
-} from "../../..";
+import { DataGridFeatures, OnyxDataGrid, type ColumnConfig } from "../../..";
 
 type TEntry = {
   id: number;
@@ -22,7 +17,7 @@ const data: TEntry[] = [
   { id: 5, name: "John", age: 42, birthday: new Date("1997-04-18") },
 ];
 
-const columns: ColumnConfig<TEntry, ColumnGroupConfig, never>[] = [
+const columns: ColumnConfig<TEntry>[] = [
   { key: "name", label: "Name" },
   { key: "age", label: "Rank" },
   { key: "birthday", label: "Birthday", type: "date" },