+++ packages/sit-onyx/src/components/OnyxDataGrid/features/resizing/resizing.ts
@@ -77,7 +77,8 @@ export const useResizing = createFeature(
         };
       });
 
-      if (showLastCol.value) columns.push({ key: EMPTY_COLUMN, type: EMPTY_COLUMN, label: "" });
+      if (showLastCol.value)
+        columns.push({ key: EMPTY_COLUMN, type: { name: EMPTY_COLUMN }, label: "" });
 
       return columns;
     };