+++ packages/sit-onyx/src/components/OnyxDrawer/types.ts
@@ -6,4 +6,10 @@ export type OnyxDrawerProps = OnyxModalDialogProps & {
    * How to align the drawer horizontally.
    */
   alignment?: Exclude<DialogAlignment, "center">;
+  /**
+   * Whether the drawer is a modal (with backdrop). If set to `false`, the backdrop will be hidden.
+   *
+   * @default true
+   */
+  modal?: boolean;
 };